Feature:
added GameManager and finish line. when the player reaches the finish line it will shift the camera and all players over. however, there is a weird bug where the players can move after they are transported. It happens when theres a camera2D change of any kind, not sure why